* commit '07ac642775b8b56d135b85e543e7206d35d49b46':
  NEXT-1019 - Rule validation
  NEXT-1019 - Add Rule Validator

+++ src/Core/Framework/DataAbstractionLayer/Write/Validation/ConstraintBuilder.php
@@ -64,7 +64,7 @@ class ConstraintBuilder
      */
     public function isFloat(): self
     {
-        $this->addConstraint(new Type('float'));
+        $this->addConstraint(new Type('numeric'));
 
         return $this;
     }
